Reorganized Hyperopt
