+++ ludwig/constants.py
@@ -86,3 +86,5 @@ FULL = 'full'
 HYPEROPT = 'hyperopt'
 STRATEGY = 'strategy'
 EXECUTOR = 'executor'
+MINIMIZE = 'minimize'
+MAXIMIZE = 'maximize'
